requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
python-dotenv>=1.0
gunicorn>=21.2
python-decouple==3.8